Updates in response to review comments on the pull request.
Renamed assertAttributeNever to assertAttributeContinuallyNotEqualTo (https://github.com/apache/incubator-brooklyn/pull/994#discussion_r43643555).

Updates as noted to deserializingClassRenames.properties (https://github.com/apache/incubator-brooklyn/pull/994#discussion-diff-43643649)

Fix package name in comment in EntityTestUtils (https://github.com/apache/incubator-brooklyn/pull/994#discussion-diff-43643773)

Removed a TODO in EntityTestUtils (https://github.com/apache/incubator-brooklyn/pull/994#discussion-diff-43643799)

Renamed HttpUtils to HttpAsserts; moved methods that weren't assertions to HttpTool, which involved bringing it out of core and into the utils module (https://github.com/apache/incubator-brooklyn/pull/994#discussion_r43644119)
