+++ core/src/main/java/org/apache/brooklyn/feed/http/HttpPollValue.java
@@ -21,7 +21,7 @@ package org.apache.brooklyn.feed.http;
 import java.util.List;
 import java.util.Map;
 
-import org.apache.brooklyn.util.core.http.HttpToolResponse;
+import org.apache.brooklyn.util.http.HttpToolResponse;
 
 /** @deprecated since 0.7.0, use {@link HttpToolResponse}.
  * the old {@link HttpPollValue} concrete class has been renamed {@link HttpToolResponse}
