+++ core/src/main/resources/org/apache/brooklyn/core/mgmt/persist/deserializingClassRenames.properties
@@ -609,8 +609,8 @@ brooklyn.util.flags.FlagUtils                                                   
 brooklyn.util.flags.MethodCoercions                                              : org.apache.brooklyn.util.core.flags.MethodCoercions
 brooklyn.util.flags.SetFromFlag                                                  : org.apache.brooklyn.util.core.flags.SetFromFlag
 brooklyn.util.flags.TypeCoercions                                                : org.apache.brooklyn.util.core.flags.TypeCoercions
-brooklyn.util.http.HttpTool                                                      : org.apache.brooklyn.util.core.http.HttpTool
-brooklyn.util.http.HttpToolResponse                                              : org.apache.brooklyn.util.core.http.HttpToolResponse
+brooklyn.util.http.HttpTool                                                      : org.apache.brooklyn.util.http.HttpTool
+brooklyn.util.http.HttpToolResponse                                              : org.apache.brooklyn.util.http.HttpToolResponse
 brooklyn.util.internal.ConfigKeySelfExtracting                                   : org.apache.brooklyn.util.core.internal.ConfigKeySelfExtracting
 brooklyn.util.internal.ssh.BackoffLimitedRetryHandler                            : org.apache.brooklyn.util.core.internal.ssh.BackoffLimitedRetryHandler
 brooklyn.util.internal.ssh.ShellAbstractTool                                     : org.apache.brooklyn.util.core.internal.ssh.ShellAbstractTool
@@ -1417,3 +1417,5 @@ brooklyn.util.internal.Repeater                                                 
 brooklyn.entity.nosql.redis.RedisClusterImpl                                     : org.apache.brooklyn.entity.nosql.redis.RedisClusterImpl
 brooklyn.test.TrustingSslSocketFactory                                           : org.apache.brooklyn.util.http.TrustingSslSocketFactory
 brooklyn.util.crypto.TrustingSslSocketFactory                                    : org.apache.brooklyn.util.crypto.TrustingSslSocketFactory
+
+org.apache.brooklyn.test.TrustingSslSocketFactory                                : org.apache.brooklyn.util.http.TrustingSslSocketFactory
