+++ usage/test-support/src/main/java/org/apache/brooklyn/test/EntityTestUtils.java
@@ -48,12 +48,11 @@ import com.google.common.collect.Sets;
 /**
  * A utility class containing tests on Entities.
  *
- * @deprecated Prefer core assertions org.apache.brooklyn.entity.EntityAsserts.
+ * @deprecated Prefer core assertions class org.apache.brooklyn.core.entity.EntityAsserts.
  */
 @Deprecated
 public class EntityTestUtils {
 
-    // TODO would be nice to have this... perhaps moving this class, or perhaps this whole project, to core/src/test ?
 //    public static LocalManagementContext newManagementContext() { return new LocalManagementContextForTests(); }
     
     // TODO Delete methods from TestUtils, to just have them here (or switch so TestUtils delegates here,
